Updated backend code with additional params and connected TakeMoney to backend
